V0.3 Heb DTO's en Entities gerefactored want die stonden verkeerd om
