+++ src/main/java/io/swagger/api/TransactionsApi.java
@@ -5,7 +5,7 @@
  */
 package io.swagger.api;
 
-import io.swagger.api.model.DTO.TransactionDTO;
+import io.swagger.api.model.Entity.Transaction;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -34,34 +34,34 @@ public interface TransactionsApi {
     @Operation(summary = "Get a specific transaction", description = "When you are an admin you are able to see all transactions. When you are a user you can only see your own transactions.", security = {
         @SecurityRequirement(name = "bearerAuth")    }, tags={ "users", "employees" })
     @ApiResponses(value = { 
-        @ApiResponse(responseCode = "200", description = "Getting transaction successful", content = @Content(mediaType = "application/json", schema = @Schema(implementation = TransactionDTO.class))),
+        @ApiResponse(responseCode = "200", description = "Getting transaction successful", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Transaction.class))),
         
         @ApiResponse(responseCode = "400", description = "bad input parameter"),
         
-        @ApiResponse(responseCode = "404", description = "Transaction not found") })
+        @ApiResponse(responseCode = "404", description = "TransactionDTO not found") })
     @RequestMapping(value = "/transactions/{id}",
         produces = { "application/json" }, 
         method = RequestMethod.GET)
-    ResponseEntity<TransactionDTO> getSpecificTransaction(@Parameter(in = ParameterIn.PATH, description = "ID from the transaction", required=true, schema=@Schema()) @PathVariable("id") Integer id);
+    ResponseEntity<Transaction> getSpecificTransaction(@Parameter(in = ParameterIn.PATH, description = "ID from the transaction", required=true, schema=@Schema()) @PathVariable("id") Integer id);
 
 
     @Operation(summary = "Gets transactions", description = "Employee can see all transactions, regular user can see only their own transactions", security = {
         @SecurityRequirement(name = "bearerAuth")    }, tags={ "users", "employees" })
     @ApiResponses(value = { 
-        @ApiResponse(responseCode = "200", description = "Succesfully retrieved transactions!", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TransactionDTO.class)))),
+        @ApiResponse(responseCode = "200", description = "Succesfully retrieved transactions!", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Transaction.class)))),
         
         @ApiResponse(responseCode = "400", description = "BadRequest") })
     @RequestMapping(value = "/transactions",
         produces = { "application/json" }, 
         method = RequestMethod.GET)
-    ResponseEntity<List<TransactionDTO>> transactionsGet(@Min(1) @Max(50) @Parameter(in = ParameterIn.QUERY, description = "The numbers of transactions to return." ,schema=@Schema(allowableValues={  }, minimum="1", maximum="50"
+    ResponseEntity<List<Transaction>> transactionsGet(@Min(1) @Max(50) @Parameter(in = ParameterIn.QUERY, description = "The numbers of transactions to return." ,schema=@Schema(allowableValues={  }, minimum="1", maximum="50"
 , defaultValue="20")) @Valid @RequestParam(value = "limit", required = false, defaultValue="20") Integer limit);
 
 
     @Operation(summary = "Create a new transaction", description = "", security = {
         @SecurityRequirement(name = "bearerAuth")    }, tags={ "users", "employees" })
     @ApiResponses(value = { 
-        @ApiResponse(responseCode = "200", description = "Created a new transaction", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TransactionDTO.class)))),
+        @ApiResponse(responseCode = "200", description = "Created a new transaction", content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Transaction.class)))),
         
         @ApiResponse(responseCode = "400", description = "bad input parameter"),
         
@@ -70,7 +70,7 @@ public interface TransactionsApi {
         produces = { "application/json" }, 
         consumes = { "application/json" }, 
         method = RequestMethod.POST)
-    ResponseEntity<List<TransactionDTO>> transactionsPost(@Parameter(in = ParameterIn.DEFAULT, description = "Create transactions", schema=@Schema()) @Valid @RequestBody TransactionDTO body);
+    ResponseEntity<List<Transaction>> transactionsPost(@Parameter(in = ParameterIn.DEFAULT, description = "Create transactions", schema=@Schema()) @Valid @RequestBody Transaction body);
 
 }
