+++ src/main/java/io/swagger/api/model/DTO/Role.java
@@ -1,0 +1,11 @@
+package io.swagger.api.model.DTO;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum Role implements GrantedAuthority {
+    ROLE_ADMIN,
+    ROLE_USER;
+    public String getAuthority() {
+        return name();
+    }
+}
