+++ src/main/java/io/swagger/api/model/DTO/TransactionDTO.java
@@ -1,265 +1,124 @@
 package io.swagger.api.model.DTO;
 
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.v3.oas.annotations.media.Schema;
-import java.math.BigDecimal;
-import org.springframework.validation.annotation.Validated;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
 
-/**
- * TransactionDTO
- */
-@Validated
-@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-05-04T11:53:18.205Z[GMT]")
+import javax.persistence.Entity;
+import javax.persistence.Id;
 
+@Entity
+public class TransactionDTO {
+    @Id
+    private Integer transactionId;
 
-public class TransactionDTO   {
-  @JsonProperty("transactionId")
-  private Integer transactionId = null;
+    public enum TypeEnum {
+        TRANSACTION("TransactionDTO"),
 
-  /**
-   * Gets or Sets type
-   */
-  public enum TypeEnum {
-    TRANSACTION("Transaction"),
-    
-    DEPOSIT("Deposit"),
-    
-    WITHDRAW("Withdraw");
+        DEPOSIT("Deposit"),
 
-    private String value;
+        WITHDRAW("Withdraw");
 
-    TypeEnum(String value) {
-      this.value = value;
-    }
+        private String value;
 
-    @Override
-    @JsonValue
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static TypeEnum fromValue(String text) {
-      for (TypeEnum b : TypeEnum.values()) {
-        if (String.valueOf(b.value).equals(text)) {
-          return b;
+        TypeEnum(String value) {
+            this.value = value;
         }
-      }
-      return null;
+
+        @Override
+        @JsonValue
+        public String toString() {
+            return String.valueOf(value);
+        }
+
+        @JsonCreator
+        public static TransactionDTO.TypeEnum fromValue(String text) {
+            for (TransactionDTO.TypeEnum b : TransactionDTO.TypeEnum.values()) {
+                if (String.valueOf(b.value).equals(text)) {
+                    return b;
+                }
+            }
+            return null;
+        }
     }
-  }
-  @JsonProperty("type")
-  private TypeEnum type = null;
 
-  @JsonProperty("timestamp")
-  private String timestamp = null;
+    private TypeEnum type;
 
-  @JsonProperty("account_from")
-  private String accountFrom = null;
+    private String timestamp;
 
-  @JsonProperty("account_to")
-  private String accountTo = null;
+    private String accountFrom;
 
-  @JsonProperty("performed_by")
-  private Integer performedBy = null;
+    private String accountTo;
 
-  @JsonProperty("amount")
-  private BigDecimal amount = null;
+    private Integer performedBy;
 
-  public TransactionDTO transactionId(Integer transactionId) {
-    this.transactionId = transactionId;
-    return this;
-  }
-
-  /**
-   * Get transactionId
-   * @return transactionId
-   **/
-  @Schema(example = "1", required = true, description = "")
-      @NotNull
+    private Double amount;
 
     public Integer getTransactionId() {
-    return transactionId;
-  }
+        return transactionId;
+    }
 
-  public void setTransactionId(Integer transactionId) {
-    this.transactionId = transactionId;
-  }
-
-  public TransactionDTO type(TypeEnum type) {
-    this.type = type;
-    return this;
-  }
-
-  /**
-   * Get type
-   * @return type
-   **/
-  @Schema(example = "Deposit", required = true, description = "")
-      @NotNull
+    public void setTransactionId(Integer transactionId) {
+        this.transactionId = transactionId;
+    }
 
     public TypeEnum getType() {
-    return type;
-  }
+        return type;
+    }
 
-  public void setType(TypeEnum type) {
-    this.type = type;
-  }
-
-  public TransactionDTO timestamp(String timestamp) {
-    this.timestamp = timestamp;
-    return this;
-  }
-
-  /**
-   * Get timestamp
-   * @return timestamp
-   **/
-  @Schema(example = "29/04/2021", required = true, description = "")
-      @NotNull
+    public void setType(TypeEnum type) {
+        this.type = type;
+    }
 
     public String getTimestamp() {
-    return timestamp;
-  }
+        return timestamp;
+    }
 
-  public void setTimestamp(String timestamp) {
-    this.timestamp = timestamp;
-  }
-
-  public TransactionDTO accountFrom(String accountFrom) {
-    this.accountFrom = accountFrom;
-    return this;
-  }
-
-  /**
-   * Get accountFrom
-   * @return accountFrom
-   **/
-  @Schema(example = "NL20INHO0032076001", required = true, description = "")
-      @NotNull
+    public void setTimestamp(String timestamp) {
+        this.timestamp = timestamp;
+    }
 
     public String getAccountFrom() {
-    return accountFrom;
-  }
+        return accountFrom;
+    }
 
-  public void setAccountFrom(String accountFrom) {
-    this.accountFrom = accountFrom;
-  }
-
-  public TransactionDTO accountTo(String accountTo) {
-    this.accountTo = accountTo;
-    return this;
-  }
-
-  /**
-   * Get accountTo
-   * @return accountTo
-   **/
-  @Schema(example = "NL20INHO0032070023", required = true, description = "")
-      @NotNull
+    public void setAccountFrom(String accountFrom) {
+        this.accountFrom = accountFrom;
+    }
 
     public String getAccountTo() {
-    return accountTo;
-  }
+        return accountTo;
+    }
 
-  public void setAccountTo(String accountTo) {
-    this.accountTo = accountTo;
-  }
-
-  public TransactionDTO performedBy(Integer performedBy) {
-    this.performedBy = performedBy;
-    return this;
-  }
-
-  /**
-   * Get performedBy
-   * @return performedBy
-   **/
-  @Schema(example = "151", required = true, description = "")
-      @NotNull
+    public void setAccountTo(String accountTo) {
+        this.accountTo = accountTo;
+    }
 
     public Integer getPerformedBy() {
-    return performedBy;
-  }
-
-  public void setPerformedBy(Integer performedBy) {
-    this.performedBy = performedBy;
-  }
-
-  public TransactionDTO amount(BigDecimal amount) {
-    this.amount = amount;
-    return this;
-  }
-
-  /**
-   * Get amount
-   * @return amount
-   **/
-  @Schema(example = "200", required = true, description = "")
-      @NotNull
-
-    @Valid
-    public BigDecimal getAmount() {
-    return amount;
-  }
-
-  public void setAmount(BigDecimal amount) {
-    this.amount = amount;
-  }
-
-
-  @Override
-  public boolean equals(java.lang.Object o) {
-    if (this == o) {
-      return true;
+        return performedBy;
     }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
+
+    public void setPerformedBy(Integer performedBy) {
+        this.performedBy = performedBy;
     }
-    TransactionDTO transactionDTO = (TransactionDTO) o;
-    return Objects.equals(this.transactionId, transactionDTO.transactionId) &&
-        Objects.equals(this.type, transactionDTO.type) &&
-        Objects.equals(this.timestamp, transactionDTO.timestamp) &&
-        Objects.equals(this.accountFrom, transactionDTO.accountFrom) &&
-        Objects.equals(this.accountTo, transactionDTO.accountTo) &&
-        Objects.equals(this.performedBy, transactionDTO.performedBy) &&
-        Objects.equals(this.amount, transactionDTO.amount);
-  }
 
-  @Override
-  public int hashCode() {
-    return Objects.hash(transactionId, type, timestamp, accountFrom, accountTo, performedBy, amount);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class TransactionDTO {\n");
-    
-    sb.append("    transactionId: ").append(toIndentedString(transactionId)).append("\n");
-    sb.append("    type: ").append(toIndentedString(type)).append("\n");
-    sb.append("    timestamp: ").append(toIndentedString(timestamp)).append("\n");
-    sb.append("    accountFrom: ").append(toIndentedString(accountFrom)).append("\n");
-    sb.append("    accountTo: ").append(toIndentedString(accountTo)).append("\n");
-    sb.append("    performedBy: ").append(toIndentedString(performedBy)).append("\n");
-    sb.append("    amount: ").append(toIndentedString(amount)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(java.lang.Object o) {
-    if (o == null) {
-      return "null";
+    public Double getAmount() {
+        return amount;
     }
-    return o.toString().replace("\n", "\n    ");
-  }
+
+    public void setAmount(Double amount) {
+        this.amount = amount;
+    }
+
+    public TransactionDTO() {
+    }
+
+    public TransactionDTO(Integer transactionId, TypeEnum type, String timestamp, String accountFrom, String accountTo, Integer performedBy, Double amount) {
+        this.transactionId = transactionId;
+        this.type = type;
+        this.timestamp = timestamp;
+        this.accountFrom = accountFrom;
+        this.accountTo = accountTo;
+        this.performedBy = performedBy;
+        this.amount = amount;
+    }
 }
