+++ src/main/java/io/swagger/api/model/Entity/Account.java
@@ -1,47 +1,187 @@
 package io.swagger.api.model.Entity;
 
+import java.util.Objects;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.api.model.DTO.AccountDTO;
+import io.swagger.v3.oas.annotations.media.Schema;
+import org.springframework.validation.annotation.Validated;
 
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import java.util.UUID;
+import javax.validation.constraints.*;
 
-@Entity
+/**
+ * Account
+ */
+@Validated
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-05-04T11:53:18.205Z[GMT]")
+
+
 public class Account {
-    @Id
-    private String IBAN;
+  @JsonProperty("IBAN")
+  private String IBAN = null;
 
-    private enum AccountTypeEnum {
-        CURRENT("Current"),
+  /**
+   * Gets or Sets accountType
+   */
+  public enum AccountTypeEnum {
+    CURRENT("Current"),
+    
+    SAVINGS("Savings");
 
-        SAVINGS("Savings");
+    private String value;
 
-        private String value;
-
-        AccountTypeEnum(String value) {
-            this.value = value;
-        }
-
-        @Override
-        @JsonValue
-        public String toString() {
-            return String.valueOf(value);
-        }
-
-        @JsonCreator
-        public static Account.AccountTypeEnum fromValue(String text) {
-            for (Account.AccountTypeEnum b : Account.AccountTypeEnum.values()) {
-                if (String.valueOf(b.value).equals(text)) {
-                    return b;
-                }
-            }
-            return null;
-        }
+    AccountTypeEnum(String value) {
+      this.value = value;
     }
 
-    private double balance;
+    @Override
+    @JsonValue
+    public String toString() {
+      return String.valueOf(value);
+    }
 
-    private UUID userID;
+    @JsonCreator
+    public static AccountTypeEnum fromValue(String text) {
+      for (AccountTypeEnum b : AccountTypeEnum.values()) {
+        if (String.valueOf(b.value).equals(text)) {
+          return b;
+        }
+      }
+      return null;
+    }
+  }
+  @JsonProperty("accountType")
+  private AccountTypeEnum accountType = null;
+
+  @JsonProperty("balance")
+  private Double balance = null;
+
+  @JsonProperty("absoluteLimit")
+  private Double absoluteLimit = null;
+
+  public Account IBAN(String IBAN) {
+    this.IBAN = IBAN;
+    return this;
+  }
+
+  /**
+   * Get IBAN
+   * @return IBAN
+   **/
+  @Schema(example = "NL91 ABNA 0417 1643 00", required = true, description = "")
+      @NotNull
+
+    public String getIBAN() {
+    return IBAN;
+  }
+
+  public void setIBAN(String IBAN) {
+    this.IBAN = IBAN;
+  }
+
+  public Account accountType(AccountTypeEnum accountType) {
+    this.accountType = accountType;
+    return this;
+  }
+
+  /**
+   * Get accountType
+   * @return accountType
+   **/
+  @Schema(example = "Current", required = true, description = "")
+      @NotNull
+
+    public AccountTypeEnum getAccountType() {
+    return accountType;
+  }
+
+  public void setAccountType(AccountTypeEnum accountType) {
+    this.accountType = accountType;
+  }
+
+  public Account balance(Double balance) {
+    this.balance = balance;
+    return this;
+  }
+
+  /**
+   * Get balance
+   * minimum: -999
+   * @return balance
+   **/
+  @Schema(example = "3242.11", required = true, description = "")
+      @NotNull
+
+  @DecimalMin("-999")  public Double getBalance() {
+    return balance;
+  }
+
+  public void setBalance(Double balance) {
+    this.balance = balance;
+  }
+
+  public Account absoluteLimit(Double absoluteLimit) {
+    this.absoluteLimit = absoluteLimit;
+    return this;
+  }
+
+  /**
+   * Get absoluteLimit
+   * @return absoluteLimit
+   **/
+  @Schema(example = "-500", required = true, description = "")
+      @NotNull
+
+    public Double getAbsoluteLimit() {
+    return absoluteLimit;
+  }
+
+  public void setAbsoluteLimit(Double absoluteLimit) {
+    this.absoluteLimit = absoluteLimit;
+  }
+
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    Account account = (Account) o;
+    return Objects.equals(this.IBAN, account.IBAN) &&
+        Objects.equals(this.accountType, account.accountType) &&
+        Objects.equals(this.balance, account.balance) &&
+        Objects.equals(this.absoluteLimit, account.absoluteLimit);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(IBAN, accountType, balance, absoluteLimit);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class Account {\n");
+    
+    sb.append("    IBAN: ").append(toIndentedString(IBAN)).append("\n");
+    sb.append("    accountType: ").append(toIndentedString(accountType)).append("\n");
+    sb.append("    balance: ").append(toIndentedString(balance)).append("\n");
+    sb.append("    absoluteLimit: ").append(toIndentedString(absoluteLimit)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
 }
