+++ src/main/java/io/swagger/api/model/Entity/User.java
@@ -1,116 +1,425 @@
 package io.swagger.api.model.Entity;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonValue;
-import io.swagger.api.model.DTO.UserDTO;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
-@Entity
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+import io.swagger.v3.oas.annotations.media.Schema;
+import org.threeten.bp.LocalDate;
+import org.springframework.validation.annotation.Validated;
+
+import javax.persistence.Id;
+import javax.validation.Valid;
+import javax.validation.constraints.*;
+
+/**
+ * User
+ */
+@Validated
+@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2022-05-04T11:53:18.205Z[GMT]")
+
+
 public class User {
-    @Id
-    private UUID userID;
+  @Id
+  @JsonProperty("userId")
+  private UUID userId = null;
 
-    private String firstName;
+  @JsonProperty("firstName")
+  private String firstName = null;
 
-    private String lastName;
+  @JsonProperty("lastName")
+  private String lastName = null;
 
-    private String email;
+  @JsonProperty("email")
+  private String email = null;
 
-    private String address;
+  @JsonProperty("dateOfBirth")
+  private LocalDate dateOfBirth = null;
 
-    private String phoneNumber;
+  @JsonProperty("address")
+  private Adress address = null;
 
-    public enum role {
-        CUSTOMER("Customer"),
+  @JsonProperty("phoneNumber")
+  private String phoneNumber = null;
 
-        EMPLOYEE("Employee");
+  /**
+   * Gets or Sets role
+   */
+  public enum RoleEnum {
+    CUSTOMER("Customer"),
+    
+    EMPLOYEE("Employee");
 
-        private String value;
+    private String value;
 
-        role(String value) {
-            this.value = value;
+    RoleEnum(String value) {
+      this.value = value;
+    }
+
+    @Override
+    @JsonValue
+    public String toString() {
+      return String.valueOf(value);
+    }
+
+    @JsonCreator
+    public static RoleEnum fromValue(String text) {
+      for (RoleEnum b : RoleEnum.values()) {
+        if (String.valueOf(b.value).equals(text)) {
+          return b;
         }
+      }
+      return null;
+    }
+  }
+  @JsonProperty("role")
+  private RoleEnum role = null;
 
-        @Override
-        @JsonValue
-        public String toString() {
-            return String.valueOf(value);
+  @JsonProperty("accounts")
+  private Account accounts = null;
+
+  /**
+   * Gets or Sets accountStatus
+   */
+  public enum AccountStatusEnum {
+    ACTIVE("Active"),
+    
+    INACTIVE("Inactive");
+
+    private String value;
+
+    AccountStatusEnum(String value) {
+      this.value = value;
+    }
+
+    @Override
+    @JsonValue
+    public String toString() {
+      return String.valueOf(value);
+    }
+
+    @JsonCreator
+    public static AccountStatusEnum fromValue(String text) {
+      for (AccountStatusEnum b : AccountStatusEnum.values()) {
+        if (String.valueOf(b.value).equals(text)) {
+          return b;
         }
-
-        @JsonCreator
-        public static User.role fromValue(String text) {
-            for (User.role b : User.role.values()) {
-                if (String.valueOf(b.value).equals(text)) {
-                    return b;
-                }
-            }
-            return null;
-        }
+      }
+      return null;
     }
-    @JsonProperty("accountList")
-    @JsonManagedReference
-    @OneToMany(mappedBy = "userID")
-    private List<Account> accountList;
+  }
+  @JsonProperty("accountStatus")
+  private AccountStatusEnum accountStatus = null;
 
-    private Double dayLimit = null;
+  @JsonProperty("dayLimit")
+  private Double dayLimit = null;
 
-    private Double transactionLimit = null;
+  @JsonProperty("transactionLimit")
+  private Double transactionLimit = null;
 
-    public User() {
-    }
+  public User userId(UUID userId) {
+    this.userId = userId;
+    return this;
+  }
 
-    public User(UUID userID, String firstName, String lastName, String email, String address, String phoneNumber, List<Account> accountList, Double dayLimit, Double transactionLimit) {
-        this.userID = userID;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.email = email;
-        this.address = address;
-        this.phoneNumber = phoneNumber;
-        this.accountList = accountList;
-        this.dayLimit = dayLimit;
-        this.transactionLimit = transactionLimit;
-    }
+  /**
+   * Get userId
+   * @return userId
+   **/
+  @Schema(example = "10", required = true, description = "")
+      @NotNull
 
-    public UUID getUserID() {
-        return userID;
-    }
+    public UUID getUserId() {
+    return userId;
+  }
+
+  public void setUserId(UUID userId) {
+    this.userId = userId;
+  }
+
+  public User firstName(String firstName) {
+    this.firstName = firstName;
+    return this;
+  }
+
+  /**
+   * Get firstName
+   * @return firstName
+   **/
+  @Schema(example = "Peter", required = true, description = "")
+      @NotNull
 
     public String getFirstName() {
-        return firstName;
-    }
+    return firstName;
+  }
+
+  public void setFirstName(String firstName) {
+    this.firstName = firstName;
+  }
+
+  public User lastName(String lastName) {
+    this.lastName = lastName;
+    return this;
+  }
+
+  /**
+   * Get lastName
+   * @return lastName
+   **/
+  @Schema(example = "Visser", required = true, description = "")
+      @NotNull
 
     public String getLastName() {
-        return lastName;
-    }
+    return lastName;
+  }
+
+  public void setLastName(String lastName) {
+    this.lastName = lastName;
+  }
+
+  public User email(String email) {
+    this.email = email;
+    return this;
+  }
+
+  /**
+   * Get email
+   * @return email
+   **/
+  @Schema(example = "Peter_Visser@hotmail.com", required = true, description = "")
+      @NotNull
 
     public String getEmail() {
-        return email;
-    }
+    return email;
+  }
 
-    public String getAddress() {
-        return address;
-    }
+  public void setEmail(String email) {
+    this.email = email;
+  }
+
+  public User dateOfBirth(LocalDate dateOfBirth) {
+    this.dateOfBirth = dateOfBirth;
+    return this;
+  }
+
+  /**
+   * Get dateOfBirth
+   * @return dateOfBirth
+   **/
+  @Schema(example = "Tue May 23 00:00:00 GMT 13", required = true, description = "")
+      @NotNull
+
+    @Valid
+    public LocalDate getDateOfBirth() {
+    return dateOfBirth;
+  }
+
+  public void setDateOfBirth(LocalDate dateOfBirth) {
+    this.dateOfBirth = dateOfBirth;
+  }
+
+  public User address(Adress address) {
+    this.address = address;
+    return this;
+  }
+
+  /**
+   * Get address
+   * @return address
+   **/
+  @Schema(required = true, description = "")
+      @NotNull
+
+    @Valid
+    public Adress getAddress() {
+    return address;
+  }
+
+  public void setAddress(Adress address) {
+    this.address = address;
+  }
+
+  public User phoneNumber(String phoneNumber) {
+    this.phoneNumber = phoneNumber;
+    return this;
+  }
+
+  /**
+   * Get phoneNumber
+   * @return phoneNumber
+   **/
+  @Schema(example = "683726482", required = true, description = "")
+      @NotNull
 
     public String getPhoneNumber() {
-        return phoneNumber;
-    }
+    return phoneNumber;
+  }
 
-    public List<Account> getAccountList() {
-        return accountList;
-    }
+  public void setPhoneNumber(String phoneNumber) {
+    this.phoneNumber = phoneNumber;
+  }
+
+  public User role(RoleEnum role) {
+    this.role = role;
+    return this;
+  }
+
+  /**
+   * Get role
+   * @return role
+   **/
+  @Schema(example = "Customer", required = true, description = "")
+      @NotNull
+
+    public RoleEnum getRole() {
+    return role;
+  }
+
+  public void setRole(RoleEnum role) {
+    this.role = role;
+  }
+
+  public User accounts(Account accounts) {
+    this.accounts = accounts;
+    return this;
+  }
+
+  /**
+   * Get accounts
+   * @return accounts
+   **/
+  @Schema(required = true, description = "")
+      @NotNull
+
+    @Valid
+    public Account getAccounts() {
+    return accounts;
+  }
+
+  public void setAccounts(Account accounts) {
+    this.accounts = accounts;
+  }
+
+  public User accountStatus(AccountStatusEnum accountStatus) {
+    this.accountStatus = accountStatus;
+    return this;
+  }
+
+  /**
+   * Get accountStatus
+   * @return accountStatus
+   **/
+  @Schema(example = "Active", required = true, description = "")
+      @NotNull
+
+    public AccountStatusEnum getAccountStatus() {
+    return accountStatus;
+  }
+
+  public void setAccountStatus(AccountStatusEnum accountStatus) {
+    this.accountStatus = accountStatus;
+  }
+
+  public User dayLimit(Double dayLimit) {
+    this.dayLimit = dayLimit;
+    return this;
+  }
+
+  /**
+   * Get dayLimit
+   * @return dayLimit
+   **/
+  @Schema(example = "200", required = true, description = "")
+      @NotNull
 
     public Double getDayLimit() {
-        return dayLimit;
-    }
+    return dayLimit;
+  }
+
+  public void setDayLimit(Double dayLimit) {
+    this.dayLimit = dayLimit;
+  }
+
+  public User transactionLimit(Double transactionLimit) {
+    this.transactionLimit = transactionLimit;
+    return this;
+  }
+
+  /**
+   * Get transactionLimit
+   * @return transactionLimit
+   **/
+  @Schema(example = "2000", required = true, description = "")
+      @NotNull
 
     public Double getTransactionLimit() {
-        return transactionLimit;
+    return transactionLimit;
+  }
+
+  public void setTransactionLimit(Double transactionLimit) {
+    this.transactionLimit = transactionLimit;
+  }
+
+
+  @Override
+  public boolean equals(java.lang.Object o) {
+    if (this == o) {
+      return true;
     }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    User user = (User) o;
+    return Objects.equals(this.userId, user.userId) &&
+        Objects.equals(this.firstName, user.firstName) &&
+        Objects.equals(this.lastName, user.lastName) &&
+        Objects.equals(this.email, user.email) &&
+        Objects.equals(this.dateOfBirth, user.dateOfBirth) &&
+        Objects.equals(this.address, user.address) &&
+        Objects.equals(this.phoneNumber, user.phoneNumber) &&
+        Objects.equals(this.role, user.role) &&
+        Objects.equals(this.accounts, user.accounts) &&
+        Objects.equals(this.accountStatus, user.accountStatus) &&
+        Objects.equals(this.dayLimit, user.dayLimit) &&
+        Objects.equals(this.transactionLimit, user.transactionLimit);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(userId, firstName, lastName, email, dateOfBirth, address, phoneNumber, role, accounts, accountStatus, dayLimit, transactionLimit);
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class User {\n");
+    
+    sb.append("    userId: ").append(toIndentedString(userId)).append("\n");
+    sb.append("    firstName: ").append(toIndentedString(firstName)).append("\n");
+    sb.append("    lastName: ").append(toIndentedString(lastName)).append("\n");
+    sb.append("    email: ").append(toIndentedString(email)).append("\n");
+    sb.append("    dateOfBirth: ").append(toIndentedString(dateOfBirth)).append("\n");
+    sb.append("    address: ").append(toIndentedString(address)).append("\n");
+    sb.append("    phoneNumber: ").append(toIndentedString(phoneNumber)).append("\n");
+    sb.append("    role: ").append(toIndentedString(role)).append("\n");
+    sb.append("    accounts: ").append(toIndentedString(accounts)).append("\n");
+    sb.append("    accountStatus: ").append(toIndentedString(accountStatus)).append("\n");
+    sb.append("    dayLimit: ").append(toIndentedString(dayLimit)).append("\n");
+    sb.append("    transactionLimit: ").append(toIndentedString(transactionLimit)).append("\n");
+    sb.append("}");
+    return sb.toString();
+  }
+
+  /**
+   * Convert the given object to string with each line indented by 4 spaces
+   * (except the first line).
+   */
+  private String toIndentedString(java.lang.Object o) {
+    if (o == null) {
+      return "null";
+    }
+    return o.toString().replace("\n", "\n    ");
+  }
 }
