+++ src/main/java/io/swagger/api/repository/AccountRepository.java
@@ -1,12 +1,11 @@
 package io.swagger.api.repository;
 
-import io.swagger.api.model.Entity.Account;
+import io.swagger.api.model.DTO.AccountDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<AccountDTO, Long> {
 
 
 }
