+++ src/main/java/io/swagger/api/repository/TransactionRepository.java
@@ -1,13 +1,10 @@
 package io.swagger.api.repository;
 
-import io.swagger.api.model.Entity.Transaction;
-import io.swagger.api.model.Entity.User;
+import io.swagger.api.model.DTO.TransactionDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<TransactionDTO, Long> {
 
 }
