+++ src/main/java/io/swagger/api/repository/UserRepository.java
@@ -1,12 +1,10 @@
 package io.swagger.api.repository;
 
-import io.swagger.api.model.Entity.User;
+import io.swagger.api.model.DTO.UserDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<UserDTO, Long> {
 
 }
